rename calc test and test data, add more test data and move check test to own file
